+++ pysys/process/monitorimpl.py
@@ -59,7 +59,7 @@ class WindowsProcessMonitor(BaseProcessMonitor):
 
 	def _getData(self, sample):
 		while True: # loop until we have both a "new" and a "last" value for CPU time
-			if self._stopping.isSet(): raise Exception('Requested to stop')
+			if self._stopping.is_set(): raise Exception('Requested to stop')
 
 			newvalues = {}
 			newvalues['time_ns'] = self._timer_ns()